+++ .github/changelog.md
@@ -1,6 +1,7 @@
 # What's New?
 
 - Hovering on recent favorite item will show how long ago it was ♥ed
+- Recently added playlist returns a max of 100 tracks, but without a cutoff period
 
 # Development
 - API documentation on /openapi